+++ acetime/zoned_date_time.go
@@ -13,8 +13,11 @@ var (
 
 // Resolved disambiguation of the ZonedDateTime from its UnixSeconds or
 // PlainDateTime.
+type ResolvedType uint8
+
 const (
-	ResolvedUnique = iota
+	ResolvedError ResolvedType = iota
+	ResolvedUnique
 	ResolvedOverlapEarlier
 	ResolvedOverlapLater
 	ResolvedGapEarlier
@@ -35,7 +38,7 @@ const (
 type ZonedDateTime struct {
 	OffsetDateTime
 	Tz       *TimeZone
-	Resolved uint8
+	Resolved ResolvedType
 }
 
 // Converts from Unix unixSeconds.
@@ -92,21 +95,6 @@ func (zdt *ZonedDateTime) Normalize(disambiguate uint8) {
 		&zdt.OffsetDateTime.PlainDateTime, disambiguate)
 	zdt.OffsetDateTime = odt
 	zdt.Resolved = resolved
-}
-
-// ZonedExtra returns the ZonedExtra object corresponding to the current
-// ZonedDateTime. This is will be always identical to the value returned by
-// ZonedExtraFromUnixSeconds().
-//
-// It will usually be identical to the value returned by
-// ZonedExtraFromPlainDateTime() except when the PlainDateTime falls in a gap
-// (FoldTypeGap). In that case, the ZonedDateTime has already been normalized
-// into a real ZonedDateTime, and ZonedExtraFromPlainDateTime() should be
-// called with the original PlainDateTime if the information about the
-// non-existent PlainDateTime is required.
-func (zdt *ZonedDateTime) ZonedExtra() ZonedExtra {
-	return zdt.Tz.findZonedExtraForUnixSeconds(
-		zdt.OffsetDateTime.UnixSeconds())
 }
 
 // String returns the given ZonedDateTime in ISO8601 format, in the form of